% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emoji_to_text.R
\name{emoji_to_text}
\alias{emoji_to_text}
\title{Convert emojis to text}
\usage{
emoji_to_text(vector, accents = FALSE, delete = FALSE)
}
\arguments{
\item{vector}{a vector of character strings.}

\item{accents}{a boolean which is TRUE if user wants to replace accents, and FALSE otherwise.}

\item{delete}{a boolean which is TRUE if user wants to delete bytes that are not matched to an emoji or accent.}
}
\value{
a vector of character strings.
}
\description{
\code{emoji_to_text} applies replaces emojis and accents with
corresponding english characters/phrases in a vector of strings.
}
\details{
This function identifies emojis and accents within a text string and replaces them
with their corresponding english characters/phrases using \code{\link{Emojis}} and \code{\link{Accents}} reference dataframes.
}
\examples{
\dontrun{
test <- c("Hi Röb, whats up 😬")
emoji_to_text(test, accents=TRUE)
#[1] "Hi Rob, whats up (grimacing face)"
}
}
\author{
Kelly Jamrog <kjamrog@wesleyan.edu>

Jacob Kuruvilla <jkuruvilla@wesleyan.edu>

Kim Pham <spham@wesleyan.edu>
}
